SCB-2094 Move etcd cache
